Use gdb.ValuePrinter tag class
GDB 14 has a "gdb.ValuePrinter" tag class that was introduced to let
GDB evolve the pretty-printer API.  Users of this tag are required to
hide any local attributes or methods.  This patch makes this change to
the Rust pretty-printers.  At present this is just a cleanup,
providing the basis for any future changes.