Expands help for error message E0161
This error message gets triggered when someone wants to move a value in
a trait object, such as a `Box<dyn Trait>`. The message gives some
advice, such as using a reference. However, this does not address the
situation where someone actually intends to move the value. Therefore,
this commit adds a suggestion for a possibility of using the `Box<Self>`
pattern where one intends to move a self value in the case of using a
boxed trait object.