Shrink `mir::Location` to 8 bytes
This type is used quite frequently and this shrinks it nicely to a
single word. I don't think we should ever have more than 4 billion MIR
statements in a single BB, that would be quite concerning.

This commit is the 1AM "just check perf" version implementation of that,
it should use IndexVec instead, I just thought it would be easier to
hack together like this (it was not).