Remap paths from other crates
Previously, `remap-path-prefix` only applied to files loaded *in the current session*. File names
loaded in a previous invocation were used unchanged. This is unfortunate for UI testing when some of
the crates involved are in the standard library (we don't want to recompile the standard library for
arbitrary tests).

Change it to apply to existing files as well. I believe this will eventually allow removing the
special-casing for `-Z simulate-remapped-rust-src-base`.